requirements.txt
pandas
numpy
seaborn
matplotlib
scipy
statsmodels
itertools
folium
scikit-learn
sklearn.linear_model
sklearn.model_selection
sklearn.metrics 
sklearn.ensemble
streamlit